There is an inconsistency in Microsoft specifications for type library minor version values. MIDL reference allows 16bit MinorVersions https://msdn.microsoft.com/en-us/library/windows/desktop/aa367306%28v=vs.85%29.aspx MSI reference allows only 8bit MinorVersions: https://msdn.microsoft.com/en-us/library/aa372092%28v=vs.85%29.aspx
The Wix Toolset follows the MSI specification.
IRESS type libraries follow the MIDL specification.

Lowering type library minor version number, or increasing the major version number is not
possible without breaking clients references to type libraries.

This workaround stops the compilation error in candle, but requires we use registry values
to register the type libraries.

WIRE reference: DOC-10308

+++ src/tools/wix/Compiler.cs
@@ -17058,7 +17058,7 @@ namespace Microsoft.Tools.WindowsInstallerXml
                             majorVersion = this.core.GetAttributeIntegerValue(sourceLineNumbers, attrib, 0, ushort.MaxValue);
                             break;
                         case "MinorVersion":
-                            minorVersion = this.core.GetAttributeIntegerValue(sourceLineNumbers, attrib, 0, byte.MaxValue);
+                            minorVersion = this.core.GetAttributeIntegerValue(sourceLineNumbers, attrib, 0, ushort.MaxValue);
                             break;
                         case "ResourceId":
                             resourceId = this.core.GetAttributeLongValue(sourceLineNumbers, attrib, int.MinValue, int.MaxValue);
